fix: migrate dependent crates to tree-sitter 0.27
tree-sitter 0.27 changes child_count() to return u32 and cargo's
links-key rule forbids two tree-sitter versions in one dependency
graph, so move-suggest, move-bounds-checker and move1-to-move2 must
move to 0.27 alongside tools-base.

+++ move-suggest/src/vector_literal.rs
@@ -195,7 +195,7 @@ fn scan_block_for_multi_push(
     suggestions: &mut Vec<Suggestion>,
     consumed: &mut Vec<(usize, usize)>,
 ) {
-    let child_count = block.child_count();
+    let child_count = block.child_count() as usize;
     let mut i: usize = 0;
 
     while i < child_count {
